Rework to a fully vectorized version
This implements a full vecrorized version of the perfect model test.
Speed up is about 400x (!!) compared to the old version without multiprocessing.
There are minor differences in the results. I think they are based on the fact
that the old version corresponds to the np.percentile kwarg
interpolate='lower' | 'higher' while the new implementation correspons to
interpolate='linear'. This needs to be tested.