Add verstion test
Test if the output changes between to versions.
Run like this
git checkout master^ OR just run it before applying changes
./run_tests.sh reference
git checkout master OR just run it after appltying changes
./run_tests.sh new

The script will check the output of the new version agains the output of the
old version and will fail if it is not equal.