Update ensemble handling
EC-EARTH (as always) is a bit problematic: there are several ensemble
members which are not available for all variables. The Filenames function
can not handle this special case. Therefore it is now fixed in the
get_filenames function. This means that I can no longer use the
Filenames function to get the modelnames and create an index of unique
model names. Therefore this functionality has also bin moved to get_filenames
and can be activated with a kwarg.