Update calculation of sigma to match description
The independence sigma calculation is updated slightly to match
the description in the paper (which should make it more straigh-
forward). The only thing thad could potentially change something
in the results in theses changes is that we now use the square
of the error to not have positive/negative errors cancel each other
out.