Update timeseries plot
Move away for the need of using config files but just give it the
results .nc file directly. There are still some information needed
from the config file. Could add them to the global attributes or
make the user supply them (would also make it more flexible but then
periods, seasons etc. which where not the target of the weighting
could be used which might not be what we want)